Fixed an authentication bug; Moved helpers into utils
