+++ base/utils/helpers.go
@@ -1,11 +1,11 @@
-package controller
+package utils
 
 import (
 	"github.com/jinzhu/gorm"
 	"github.com/pkg/errors"
 )
 
-func panicIfDBError(db *gorm.DB, message string) {
+func PanicIfDBError(db *gorm.DB, message string) {
 	if db.Error != nil {
 		panic(errors.Wrap(db.Error, message))
 	}
